Create JobFlagTrait
