+++ src/Entities/Job.php
@@ -3,16 +3,16 @@
 namespace Tatter\Workflows\Entities;
 
 use CodeIgniter\Entity\Entity;
-use CodeIgniter\I18n\Time;
 use RuntimeException;
 use Tatter\Workflows\Exceptions\WorkflowsException;
-use Tatter\Workflows\Models\JobflagModel;
 use Tatter\Workflows\Models\JobModel;
 use Tatter\Workflows\Models\StageModel;
 use Tatter\Workflows\Models\WorkflowModel;
 
 class Job extends Entity
 {
+    use JobFlagTrait;
+
     protected $dates = [
         'created_at',
         'updated_at',
@@ -41,13 +41,21 @@ class Job extends Entity
     protected ?Workflow $workflow = null;
 
     /**
-     * Stored flags from `jobflags`.
+     * Verifies the primary key to prevent operations on
+     * un-created database entries.
      *
-     * @var array<string,Time>|null
+     * @throws RuntimeException
+     *
+     * @return $this
      */
-    protected ?array $flags = null;
+    protected function ensureCreated(): self
+    {
+        if (empty($this->attributes['id'])) {
+            throw new RuntimeException('Job must be created first.');
+        }
 
-    //--------------------------------------------------------------------
+        return $this;
+    }
 
     /**
      * Returns the URL to show this Job.
@@ -58,100 +66,6 @@ class Job extends Entity
 
         return site_url($base . 'show/' . $this->attributes['id']);
     }
-
-    /**
-     * Fetches, stores, and returns all this job's flags (from `jobflags`).
-     *
-     * @return array<string,Time>
-     */
-    public function getFlags(): array
-    {
-        $this->ensureCreated();
-
-        if (null === $this->flags) {
-            $this->flags = [];
-
-            foreach (model(JobflagModel::class)->where('job_id', $this->attributes['id'])->findAll() as $flag) {
-                $this->flags[$flag->name] = new Time($flag->created_at);
-            }
-        }
-
-        return $this->flags;
-    }
-
-    /**
-     * Gets a flag by its name.
-     */
-    public function getFlag(string $name): ?Time
-    {
-        return $this->getFlags()[$name] ?? null;
-    }
-
-    /**
-     * Creates a flag for the given name.
-     *
-     * @return $this
-     */
-    public function setFlag(string $name): self
-    {
-        if (isset($this->getFlags()[$name])) {
-            model(JobflagModel::class)
-                ->where('job_id', $this->attributes['id'])
-                ->where('name', $name)
-                ->update(null, ['created_at' => date('Y-m-d H:i:s'),
-                ]);
-        } else {
-            $this->flags[$name] = new Time('now');
-
-            model(JobflagModel::class)->insert([
-                'job_id' => $this->attributes['id'],
-                'name'   => $name,
-            ]);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Removes a flag for the given name.
-     *
-     * @return $this
-     */
-    public function clearFlag(string $name): self
-    {
-        $this->ensureCreated();
-
-        model(JobflagModel::class)
-            ->where('job_id', $this->attributes['id'])
-            ->where('name', $name)
-            ->delete();
-
-        if (is_array($this->flags) && isset($this->flags[$name])) {
-            unset($this->flags[$name]);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Removes all flags.
-     *
-     * @return $this
-     */
-    public function clearFlags(): self
-    {
-        $this->ensureCreated();
-
-        model(JobflagModel::class)
-            ->where('job_id', $this->attributes['id'])
-            ->delete();
-
-        $this->flags = [];
-
-        return $this;
-    }
-
-    //--------------------------------------------------------------------
 
     /**
      * Gets the current Stage, or null for a completed Job.
@@ -301,23 +215,6 @@ class Job extends Entity
         $this->stageFlag              = false;
 
         return $results;
-    }
-
-    /**
-     * Verifies the primary key to prevent operations on
-     * un-created database entries.
-     *
-     * @throws RuntimeException
-     *
-     * @return $this
-     */
-    protected function ensureCreated(): self
-    {
-        if (empty($this->attributes['id'])) {
-            throw new RuntimeException('Job must be created first.');
-        }
-
-        return $this;
     }
 
     /**
